requirements.txt
fastapi==0.111.1
SQLAlchemy==2.0.31


pandas==2.2.2
openpyxl==3.1.2
python-docx==1.1.2
pytesseract ==0.3.13

#alembic init alembic